Create a delta model. Move package generation into the project.
