+++ psqlpack/src/model/mod.rs
@@ -1,7 +1,9 @@
 mod profiles;
 mod project;
 mod package;
+mod delta;
 
 pub use self::profiles::PublishProfile;
 pub use self::project::Project;
 pub use self::package::Package;
+pub use self::delta::Delta;
