--- psqlpack/src/operation/mod.rs
@@ -1,3 +1,0 @@
-mod package;
-
-pub use self::package::package;
